requirements.txt
paramiko>=3.5.0
stem>=1.8.3
colorama>=0.4.7
argparse>=1.4.0
PySocks>=1.7.2
tqdm>=4.65.0
requests>=2.31.0